+++ classes/class.ilSrLpReportPlugin.php
@@ -93,6 +93,25 @@ class ilSrLpReportPlugin extends ilUserInterfaceHookPlugin
                 }
                 break;
 
+            case "Services/Object":
+                switch ($a_event) {
+                    case "putObjectInTree":
+                        if (in_array(self::dic()->objDataCache()->lookupType(self::dic()->objDataCache()->lookupObjId($a_parameter["parent_ref_id"])), ilSrLpReportUIHookGUI::TYPES)) {
+                            self::reports()->syncPositionPermissionsWithChildren($a_parameter["parent_ref_id"], $a_parameter["object"]->getRefId());
+                        }
+                        break;
+
+                    case "update":
+                        if (in_array(self::dic()->objDataCache()->lookupType(self::dic()->objDataCache()->lookupObjId($a_parameter["ref_id"])), ilSrLpReportUIHookGUI::TYPES)) {
+                            self::reports()->syncPositionPermissionsWithChildrens($a_parameter["ref_id"]);
+                        }
+                        break;
+
+                    default:
+                        break;
+                }
+                break;
+
             default:
                 break;
         }